requirements.txt
python-docx
requests
beautifulsoup4
docxtpl